requirements.txt
torch
torchvision
ultralytics
opencv-python
pyyaml
scikit-learn
matplotlib
seaborn
tqdm
pandas
grad-cam
huggingface_hub
